requirements.txt
requests
redis==4.5.5
pyspark==3.5.0
apache-airflow-providers-apache-spark==5.0.0
